security: fix critical/high auth gaps and SSRF in production notification delivery
- setup/bootstrap-config: add requireWizardSession guard — previously leaked
  GITEA_ADMIN_TOKEN/USER/PASSWORD to any unauthenticated caller on unfinished setup
- tool-groups (route, [id], [id]/tools): add requireAdmin to all handlers —
  entirely unauthenticated, allowed anonymous tier manipulation and tool injection
- environments/[id]/agents: add requireAdmin to POST — unauthenticated agent linking
- environments/[id]/gateway: add requireAdmin to POST — unauthenticated forced
  gateway restart/redeploy against any environment
- agents/route: change requireAuth false → true — anonymous agent creation fed
  the orchestration pipeline
- lib/ssrf-guard.ts: extract shared isPrivateUrl/isPrivateIp into a reusable module
- lib/notifications.ts: apply SSRF guard to production webhook delivery — the
  DNS-resolution guard was only on the test endpoint, not actual event dispatch
- notification-channels/[id]/test: import from shared ssrf-guard module
