+++ apps/web/src/app/api/setup/bootstrap-config/route.ts
@@ -1,11 +1,15 @@
 export const dynamic = 'force-dynamic'
 
-import { NextResponse } from 'next/server'
+import { NextRequest, NextResponse } from 'next/server'
 import { prisma } from '@/lib/db'
+import { requireWizardSession } from '@/lib/setup-guard'
 
 // Returns server-side env vars needed by the setup wizard UI.
 // (NEXT_PUBLIC_ vars are baked at build time so we expose them server-side instead.)
-export async function GET() {
+export async function GET(req: NextRequest) {
+  if (!await requireWizardSession(req)) {
+    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 })
+  }
   // Only expose Gitea credentials during setup — after completion they're
   // stored encrypted in the database and this endpoint should not leak them.
   const setting = await prisma.systemSetting.findUnique({ where: { key: 'setup.completed' } })
